matching-service/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
pydantic
